+++ utils/reaction.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2024   Loh Wah Kiang
+ * Copyright (c) 2025   Loh Wah Kiang
  *
  * openGauss is licensed under Mulan PSL v2.
  * You can use this software according to the terms and conditions of the Mulan PSL v2.
@@ -14,25 +14,398 @@
  * -------------------------------------------------------------------------
  */
 "use strict";
+
+const { minify } = require("html-minifier-terser");
+const jsdom = require("jsdom");
+const htmlTags = require("./data/html-tags.json");
+
 /**
  * Submodule handles http responses, which are preprocessed by jsdom to manipulate the data before presenting to the client
- * @module src_reaction
+ * @module utils_reaction
  */
 module.exports = async (...args) => {
   return new Promise(async (resolve, reject) => {
     const [params, obj] = args;
     const [pathname, curdir] = params;
     const [library, sys, cosetting] = obj;
-    const { handler, getNestedObject } = library.utils;
+    const { handler, getNestedObject, sanbox } = library.utils;
     const { fs, path, logger } = sys;
-    const { minify } = require("html-minifier-terser");
-    const jsdom = require("jsdom");
 
     try {
-      const molecule = await require("./reaction/molecule")(params, obj);
-
+      const basic =
+        /\s?<!doctype html>|(<html\b[^>]*>|<body\b[^>]*>|<x-[^>]+>)+/i;
+      const full = new RegExp(
+        htmlTags.map((tag) => `<${tag}\\b[^>]*>`).join("|"),
+        "i"
+      );
       let lib = {};
       let components = { defaulturl: "" };
+
+      /**
+       * The main objective is convert css data in object type to jsdom format and append to parent
+       * @alias module:utils_reaction.import_css
+       * @param {...Object} args - 3 parameters
+       * @param {Object} args[0] - doc is an object of jsdom window.document
+       * @param {Object} args[1] - data is an object which listing css link source
+       * @param {Object} args[2] - params is an object which use to concat data object value
+       */
+      const import_css = (...args) => {
+        let [doc, data, params] = args;
+        try {
+          let el = doc.getElementsByTagName("head").item(0);
+          for (let [key, val] of Object.entries(data)) {
+            if (val.length > 0) {
+              for (let href of val) {
+                let gfgData = doc.createElement("link");
+                let url = href;
+                if (key != "other") url = params[key] + href;
+                let attributes = JSON.parse(
+                  `{"rel":"stylesheet","type":"text/css","href":"${url}"}`
+                );
+
+                Object.keys(attributes).forEach((attr) => {
+                  gfgData.setAttribute(attr, attributes[attr]);
+                });
+                el.appendChild(gfgData);
+              }
+            }
+          }
+          return;
+        } catch (error) {
+          return error;
+        }
+      };
+
+      /**
+       * The main objective is convert js data in object type to jsdom format and append to parent
+       * @alias module:utils_reaction.import_js
+       * @param {...Object} args - 3 parameters
+       * @param {Object} args[0] - doc is an object of jsdom window.document
+       * @param {Object} args[1] - data is an object which listing js link source
+       * @param {Object} args[2] - params is an object which use to concat data object value
+       */
+      const import_js = (...args) => {
+        let [doc, data, params] = args;
+        try {
+          let el = doc.getElementsByTagName("head").item(0);
+          for (let [key, val] of Object.entries(data)) {
+            if (val.length > 0) {
+              for (let href of val) {
+                let gfgData = doc.createElement("script");
+                let url = href;
+                if (key != "other") url = params[key] + href;
+                let attributes = JSON.parse(
+                  `{"type":"text/javascript","src":"${url}"}`
+                );
+
+                Object.keys(attributes).forEach((attr) => {
+                  gfgData.setAttribute(attr, attributes[attr]);
+                });
+                el.appendChild(gfgData);
+              }
+            }
+          }
+          return;
+        } catch (error) {
+          return error;
+        }
+      };
+
+      /**
+       * The main objective is concat string become complete url for ES Module
+       * @alias module:utils_reaction.import_mjs
+       * @param {...Object} args - 3 parameters
+       * @param {Object} args[0] - doc is an object of jsdom window.document
+       * @param {Object} args[1] - data is an object which listing js link source
+       * @param {Object} args[2] - params is an object which use to concat data object value
+       */
+      const import_mjs = (...args) => {
+        let [data, params] = args;
+        try {
+          let initialize = {},
+            libname = [];
+          for (let [key, val] of Object.entries(data)) {
+            if (key == "initialize") initialize[key] = val;
+            else {
+              if (val.length > 0) {
+                for (let href of val) {
+                  let url = href;
+                  if (key != "other") url = params[key] + href;
+                  libname.push(url);
+                }
+              }
+            }
+          }
+          return { initialize, lib: libname };
+        } catch (error) {
+          return error;
+        }
+      };
+
+      /**
+       * The main objective is convert less.js data in object type to jsdom format and append to parent
+       * @alias module:utils_reaction.import_less
+       * @param {...Object} args - 3 parameters
+       * @param {Object} args[0] - doc is an object of jsdom window.document
+       * @param {Object} args[1] - data is an object which listing less.js link source
+       * @param {Object} args[2] - params is an object which use to concat data object value
+       */
+      const import_less = (...args) => {
+        let [doc, data, params] = args;
+        try {
+          let el = doc.getElementsByTagName("head").item(0);
+          for (let [key, val] of Object.entries(data.style)) {
+            if (val.length > 0) {
+              for (let href of val) {
+                let gfgData = doc.createElement("link");
+                let url = href;
+                if (key != "other") url = params[key] + href;
+                let attributes = JSON.parse(
+                  `{"rel":"stylesheet/less","type":"text/css","href":"${url}"}`
+                );
+
+                Object.keys(attributes).forEach((attr) => {
+                  gfgData.setAttribute(attr, attributes[attr]);
+                });
+                el.appendChild(gfgData);
+              }
+            }
+          }
+
+          for (let [key, val] of Object.entries(data.engine)) {
+            if (val !== "") {
+              let engine = {};
+              engine[key] = [val];
+              import_js(doc, engine, params);
+            }
+          }
+
+          return;
+        } catch (error) {
+          return error;
+        }
+      };
+
+      /**
+       * The main objective is read a file content and minify become one row
+       * @alias module:utils_reaction.get_domhtml
+       * @param {...Object} args - 1 parameters
+       * @param {String} args[0] - file is file name which emebed absolute path
+       * @returns {String} - Return undefined|text
+       */
+      const get_domhtml = (...args) => {
+        let [file] = args;
+        try {
+          let output;
+          if (fs.existsSync(file)) {
+            output = fs.readFileSync(file, "utf8").trim();
+          }
+          return output;
+        } catch (error) {
+          return error;
+        }
+      };
+
+      /**
+       * Replace specific character from text base on object key name
+       * Keyword <-{name}>
+       * @alias module:reaction.str_replace
+       * @param {...Object} args - 2 parameters
+       * @param {String} args[0] - text is a statement in string value
+       * @param {Object} args[1] - params a sets of values for change
+       * @returns {String} - Return unchange or changed text
+       */
+      const str_replace = (...args) => {
+        let [text, params] = args;
+        let output = text;
+        for (let [key, val] of Object.entries(params)) {
+          let name = `<-{${key}}>`;
+          while (output.indexOf(name) > -1) {
+            let idx = output.indexOf(name);
+            output =
+              output.substring(0, idx) +
+              val +
+              output.substring(idx + name.length);
+          }
+        }
+        return output;
+      };
+
+      /**
+       * The main objective is indentify the string in html tag format
+       * https://github.com/sindresorhus/is-html
+       * @alias module:reaction.moleculde.indentify_html
+       * @param {...Object} args - 1 parameters
+       * @param {String} args[0] - buffhtml is a string for checking valid in html tag format
+       * @returns {Object} - Return valid stting | null
+       */
+      const indentify_html = (...args) => {
+        let [buffhtml] = args;
+        let output = buffhtml;
+        try {
+          let html = buffhtml.trim().slice(0, 1000);
+          let result = basic.test(html) || full.test(html);
+          if (!result) output = null;
+        } catch (error) {
+          output = null;
+        } finally {
+          return output;
+        }
+      };
+
+      /**
+       * The main objective is read a list of file content and minify become one row
+       * @alias module:reaction.moleculde.get_filenames
+       * @param {...Object} args - 1 parameters
+       * @param {Object} args[0] - node is an object provide directory path and filter list
+       * @param {Array} args[1] - included is check the file type which accept form the list
+       * @returns {Object} - Return undefined|text
+       */
+      const get_filenames = async (...args) => {
+        const [node, included = []] = args;
+        let files = await fs
+          .readdirSync(path.join(node.path))
+          .filter((filename) => {
+            let extname = path.extname(filename);
+            if (
+              extname !== "" &&
+              !included.includes(extname) &&
+              !node.excluded.includes(filename)
+            ) {
+              return filename;
+            }
+          });
+        let docs = [];
+        if (!files) return [];
+        for (let val of files) {
+          docs.push(get_domhtml(path.join(node.path, val)));
+        }
+        if (node.external) {
+          for (let val of node.external) {
+            let htmlstr = indentify_html(val);
+            if (htmlstr) docs.push(val);
+            else if (fs.existsSync(val)) docs.push(get_domhtml(val));
+          }
+        }
+        if (node.htmlstr) {
+          for (let val of node.htmlstr) {
+            let htmlstr = indentify_html(val);
+            if (htmlstr) docs.push(val);
+          }
+        }
+        return await Promise.all(docs);
+      };
+
+      /**
+       * The main objective is combine a list of file become one html text
+       * @alias module:reaction.moleculde.combine_layer
+       * @param {...Object} args - 1 parameters
+       * @param {Object} args[0] - layer is an object the list of files for merge purpose
+       * @param {Object} args[1] - params is object use for write data to relevent html element
+       * @returns {Object} - Return object
+       */
+      const combine_layer = (...args) => {
+        return new Promise(async (resolve, reject) => {
+          const [layer, params] = args;
+          const { JSDOM } = jsdom;
+          try {
+            let output = { code: 0, msg: "", data: null };
+            let master_dom, master_doc;
+            let conv;
+
+            if (indentify_html(layer.layouts)) conv = layer.layouts;
+            else conv = get_domhtml(layer.layouts);
+            let [layouts, childlists] = await Promise.all([
+              conv,
+              get_filenames(layer.childs, ["*.html"]),
+            ]);
+
+            let { message, stack } = layouts;
+            if (!message && !stack) {
+              master_dom = new JSDOM(str_replace(layouts, params));
+              master_doc = master_dom.window.document;
+
+              for (let childlist of childlists) {
+                let child_doc = new JSDOM().window.document;
+                let body = child_doc.querySelector("body");
+                body.innerHTML = str_replace(childlist, params);
+                let body_node = body.childNodes[0];
+                if (body_node && body_node.nodeName == "STATEMENT") {
+                  let statement = body
+                    .querySelector(body_node.nodeName.toLowerCase())
+                    .querySelectorAll("*");
+                  let attrname = body_node.getAttribute("name");
+                  let attraction = body_node.getAttribute("action");
+                  switch (attraction) {
+                    case "overwrite":
+                      let mel = master_doc.querySelector(attrname);
+                      let cel = body_node.querySelector(attrname);
+                      let attrs = cel.getAttributeNames();
+                      if (mel && cel) {
+                        for (let attr of attrs) {
+                          let val = cel.getAttribute(attr);
+                          mel.setAttribute(attr, val);
+                        }
+                        mel.innerHTML = cel.innerHTML;
+                      }
+                      break;
+
+                    case "append":
+                      let childNodes = master_doc.querySelector(attrname);
+                      for (const el of statement) childNodes.append(el);
+                      break;
+                  }
+                }
+              }
+
+              output.data = master_dom.serialize();
+            } else {
+              throw {
+                message: message,
+                stack: stack,
+              };
+            }
+            resolve(output);
+          } catch (error) {
+            resolve({ msg: error.message, data: error.stack });
+          }
+        });
+      };
+
+      /**
+       * The main objective is combine a list of file become one html text
+       * @alias module:reaction.moleculde.combine_layer
+       * @param {...Object} args - 1 parameters
+       * @param {Object} args[0] - layer is an object the list of files for merge purpose
+       * @param {Object} args[1] - params is object use for write data to relevent html element
+       * @returns {Object} - Return object
+       */
+      const single_layer = (...args) => {
+        return new Promise(async (resolve, reject) => {
+          const [layer, params] = args;
+          const { JSDOM } = jsdom;
+          try {
+            let output = { code: 0, msg: "", data: null };
+            let master_dom;
+            let layouts = indentify_html(layer);
+            if (!layouts) layouts = await get_domhtml(layer);
+
+            let { message, stack } = layouts;
+            if (!message && !stack) {
+              master_dom = new JSDOM(str_replace(layouts, params));
+              output.data = master_dom.serialize();
+            } else {
+              throw {
+                message: message,
+                stack: stack,
+              };
+            }
+            resolve(output);
+          } catch (error) {
+            resolve({ msg: error.message, data: error.stack });
+          }
+        });
+      };
 
       /**
        * Merge multi css file to be sinlge string and render to frontend
@@ -125,25 +498,22 @@ module.exports = async (...args) => {
                 // else if (!islayer) isvalid = path.extname(layer.layouts);
                 if (!isview) {
                   if (path.extname(view) == ".html") isvalid = true;
-                  else isvalid = molecule.indentify_html(view);
+                  else isvalid = indentify_html(view);
                 } else if (!islayer) {
                   if (path.extname(layer.layouts) == ".html") isvalid = true;
-                  else isvalid = molecule.indentify_html(layer.layouts);
+                  else isvalid = indentify_html(layer.layouts);
                 }
                 if (isvalid) {
                   if (!islayer) {
                     if (!isview) layer.childs.external.push(view);
-                    let { code, msg, data } = await molecule.combine_layer(
+                    let { code, msg, data } = await combine_layer(
                       layer,
                       params
                     );
                     if (code == 0) layouts = data;
                     else throw { msg: msg, data: data };
                   } else if (!isview) {
-                    let { code, msg, data } = await molecule.single_layer(
-                      view,
-                      params
-                    );
+                    let { code, msg, data } = await single_layer(view, params);
 
                     if (code == 0) view = data;
                     else throw { msg: msg, data: data };
@@ -169,25 +539,25 @@ module.exports = async (...args) => {
                 if (found) found.innerHTML = content;
               }
 
-              let preload = await molecule.get_domhtml(
-                path.join(pathname, "browser", "preload.html")
+              let preload = await get_domhtml(
+                path.join(pathname, "data", "preload.html")
               );
               document.querySelector("body").innerHTML += preload;
               let script = document.createElement("script");
               script.type = "text/javascript";
               script.innerHTML = `var mjs=${JSON.stringify(
-                molecule.import_mjs(mjs, params)
+                import_mjs(mjs, params)
               )};`;
               if (Object.keys(injectionjs.variables).length > 0)
                 script.innerHTML += `var injectionjs=${JSON.stringify(
                   injectionjs.variables
                 )}`;
               document.getElementsByTagName("head")[0].appendChild(script);
-              let rtnimport_css = molecule.import_css(document, css, params);
+              let rtnimport_css = import_css(document, css, params);
               if (rtnimport_css) throw rtnimport_css;
-              let rtnimport_js = molecule.import_js(document, js, params);
+              let rtnimport_js = import_js(document, js, params);
               if (rtnimport_js) throw rtnimport_js;
-              let rtnimport_less = molecule.import_less(document, less, params);
+              let rtnimport_less = import_less(document, less, params);
               if (rtnimport_less) throw rtnimport_less;
               res.status(status).send(
                 await minify(dom.serialize(), {
@@ -345,24 +715,21 @@ module.exports = async (...args) => {
        * The main purpose is to prevent users from improperly using async/await and promise
        * methods to trigger unpredictable errors and cause the entire system to shut down.
        * When server receive the request from client, will proceed
-       * @alias module:reaction.sanbox
+       * @alias module:reaction.inspector
        * @param {...Object} args - 1 parameters
        * @param {Object} args[0] - fn is mehtod/fuction for execution
        * @param {Array} args[1] - params is in array type which contant with request and response variable
        * @returns {Object} - Either return object or return data from execution function
        */
-      const sanbox = async (...args) => {
+      const inspector = async (...args) => {
         let [fn, params] = args;
         let [, response] = params;
         try {
           response.inspector = async (...args) => {
             let [fn, params] = args;
             try {
-              let result = fn.apply(null, params);
-              if (result instanceof Promise) {
-                result = await result;
-                if (result instanceof ReferenceError) throw result;
-              } else if (result instanceof ReferenceError) throw result;
+              let result = await sanbox(fn, params);
+              if (result.code) throw result;
               return result;
             } catch (error) {
               if (error.msg) response.err.error = error.msg;
@@ -372,11 +739,8 @@ module.exports = async (...args) => {
             }
           };
 
-          let result = fn.apply(null, params);
-          if (result instanceof Promise) {
-            result = await result;
-            if (result instanceof ReferenceError) throw result;
-          } else if (result instanceof ReferenceError) throw result;
+          let result = await sanbox(fn, params);
+          if (result.code) throw result;
           return result;
         } catch (error) {
           if (error.msg) response.err.error = error.msg;
@@ -385,6 +749,7 @@ module.exports = async (...args) => {
           return response;
         }
       };
+
       /**
        * The main objective is on listen register url from http client
        * @alias module:reaction.onrequest
@@ -462,11 +827,11 @@ module.exports = async (...args) => {
 
                 let queuertn;
                 if (permit && fn.idx == idx) {
-                  queuertn = sanbox(queue[fname], [orireq, response]);
+                  queuertn = inspector(queue[fname], [orireq, response]);
                   if (queuertn instanceof Promise) queuertn = await queuertn;
                   queuertn["action"] = queuertn.render;
                 } else if (fn.idx != idx) {
-                  queuertn = sanbox(queue[fname], [orireq, response]);
+                  queuertn = inspector(queue[fname], [orireq, response]);
                   if (queuertn instanceof Promise) queuertn = await queuertn;
                   queuertn["action"] = queuertn.render;
                 }
@@ -571,9 +936,8 @@ module.exports = async (...args) => {
             };
           } else {
             err["status"] = errcode;
-            if (errcode >= 500)
-              err["view"] = `${pathname}/browser/error/500.html`;
-            else err["view"] = `${pathname}/browser/error/404.html`;
+            if (errcode >= 500) err["view"] = `${pathname}/data/error/500.html`;
+            else err["view"] = `${pathname}/data/error/404.html`;
             let msg;
             if (typeof errmessage == "string") msg = errmessage;
             else msg = JSON.stringify(errmessage);
@@ -641,15 +1005,13 @@ module.exports = async (...args) => {
         }
       };
 
-      lib = {
+      resolve({
         ...lib,
-        ...{
-          get guiapi() {
-            return components;
-          },
+
+        get guiapi() {
+          return components;
         },
-      };
-      resolve(lib);
+      });
     } catch (error) {
       reject(error);
     }